Update tests, bad imports
