+++ ord_schema/updates_test.py
@@ -21,25 +21,6 @@ from ord_schema.proto import reaction_pb2
 from ord_schema.proto import dataset_pb2
 
 
-class UpdatesTest(absltest.TestCase):
-
-    def test_resolve_names(self):
-        roundtrip_smi = lambda smi: Chem.MolToSmiles(Chem.MolFromSmiles(smi))
-        message = reaction_pb2.Reaction()
-        message.inputs['test'].components.add().identifiers.add(type='NAME',
-                                                                value='aspirin')
-        self.assertTrue(updates.resolve_names(message))
-        resolved_smi = roundtrip_smi(
-            message.inputs['test'].components[0].identifiers[1].value)
-        self.assertEqual(resolved_smi, roundtrip_smi('CC(=O)Oc1ccccc1C(O)=O'))
-        self.assertEqual(
-            message.inputs['test'].components[0].identifiers[1].type,
-            reaction_pb2.CompoundIdentifier.IdentifierType.SMILES)
-        self.assertRegex(
-            message.inputs['test'].components[0].identifiers[1].details,
-            'NAME resolved')
-
-
 class UpdateReactionTest(absltest.TestCase):
 
     def test_with_updates_simple(self):
